Program 5 Description
This is a one to two page document describing teams final project. Describe the game being implemented, languages and frameworks being used, describe how the client/server relationship will be implemented, the look and feel of the user interface and anything else you would like to include. This document should have a cover page with team number, section and members and organized and written in a professional manner.